add GOOS and GOARCH to version; separate packages for helm/ansible subcommands
