+++ internal/cmd/ansible-operator/version/cmd_test.go
