+++ internal/cmd/helm-operator/version/cmd_test.go
